+++ internal/test/e2e/grpc/main.go
@@ -14,8 +14,10 @@ import (
 	"time"
 
 	"google.golang.org/grpc"
+	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/credentials/insecure"
 	pb "google.golang.org/grpc/examples/helloworld/helloworld"
+	"google.golang.org/grpc/status"
 )
 
 const port = 1701
@@ -26,6 +28,9 @@ type server struct {
 
 func (s *server) SayHello(ctx context.Context, in *pb.HelloRequest) (*pb.HelloReply, error) {
 	log.Printf("Received: %v", in.GetName())
+	if in.GetName() == "unimplemented" {
+		return nil, status.Error(codes.Unimplemented, "unimplmented")
+	}
 	return &pb.HelloReply{Message: "Hello " + in.GetName()}, nil
 }
 
@@ -67,6 +72,13 @@ func main() {
 		log.Fatalf("could not greet: %v", err)
 	}
 	log.Printf("Greeting: %s", r.GetMessage())
+
+	// Contact the server expecting a server error
+	_, err = c.SayHello(ctx, &pb.HelloRequest{Name: "unimplemented"})
+	if err == nil {
+		log.Fatalf("expected an error but none was received")
+	}
+	log.Printf("received expected error: %+v", err)
 
 	s.GracefulStop()
 	<-done